Add a 'settings' button and place the main window widgets to a separate function.
