+++ app/settings.py
@@ -1,0 +1,2 @@
+import tkinter as tk
+from tkinter import ttk
